+++ lib/paygate/transaction.rb
@@ -36,6 +36,17 @@ module Paygate
       r
     end
 
+    # Doc: https://km.paygate.net/pages/viewpage.action?pageId=9207875
+    def verify
+      params = { tid: tid, verifyNum: 100 }
+
+      uri = URI('https://service.paygate.net/djemals/settle/verifyReceived.jsp')
+      uri.query = ::URI.encode_www_form(params)
+      response = ::Net::HTTP.get_response(uri)
+
+      Response.build_from_net_http_response(:verify, response)
+    end
+
     private
 
     def self.refund_api_url